+++ Bharatiya yodha/WebContent/WEB-INF/jsp/adminbanner.jsp
@@ -215,7 +215,7 @@ Smartphone Compatible web template, free webdesigns for Nokia, Samsung, LG, Sony
 			<!-- //gallery -->
 			</td>
 			<td>
-		  <a href="${deleteadminbanner}"><i class="fa fa-trash" aria-hidden="true"></i></a>
+		  <a href="${deleteadminbanner}" onclick="return confirmActiondelete();"><i class="fa fa-trash" aria-hidden="true"></i></a>
 		  <a href="${editadminbanner}"><i class="fa fa-pencil" aria-hidden="true"></i></a>
 		  </td>
 		 
@@ -271,5 +271,15 @@ $(document).ready(function(){
 });
 
 </script> -->
+
+<script type="text/javascript">
+function confirmActiondelete() {
+	if (!confirm("Are You Sure You want to Delete?")) {
+		return false;
+	}
+}
+
+
+</script>
 </body>
 </html>
